requirements.txt
openpyxl>=3.1.0
icalendar>=5.0.0
pyyaml>=6.0
playwright
